Tenant/requirements.txt
fastapi
uvicorn
psycopg2-binary
paho-mqtt
python-dotenv